feat: add rust_ops crate for op_change
