+++ rust_ops/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "rust_ops"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["staticlib"]
