+++ src/ops_rs.h
@@ -1,0 +1,16 @@
+#ifndef OPS_RS_H
+#define OPS_RS_H
+
+#include "structs.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int rs_op_change(oparg_T *oap);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // OPS_RS_H
